Incorporate feedback from Xamarin team
