--- src/Android/Resources/layout/CipherViewCell.axml
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:minHeight="44dp"
-    android:gravity="center_vertical">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:paddingLeft="2.2dp">
-        <LinearLayout
-            android:orientation="horizontal"
-            android:layout_width="39.8dp"
-            android:layout_height="match_parent"
-            android:layout_gravity="center_vertical"
-            android:gravity="center">
-            <bit.droid.renderers.IconImageView
-                android:id="@+id/CipherCellIconImage"
-                android:layout_width="22dp"
-                android:layout_height="22dp"
-                android:layout_gravity="center"
-                android:gravity="center" />
-            <TextView
-                android:id="@+id/CipherCellIcon"
-                android:layout_width="26dp"
-                android:layout_height="26dp"
-                android:layout_gravity="center"
-                android:gravity="center"
-                android:text="[X]" />
-        </LinearLayout>
-        <LinearLayout
-            android:id="@+id/CipherCellContent"
-            android:orientation="vertical"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:layout_gravity="center_vertical"
-            android:gravity="center"
-            android:paddingVertical="7.65dp">
-            <LinearLayout
-                android:id="@+id/CipherCellContentTop"
-                android:orientation="horizontal"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-                <TextView
-                    android:id="@+id/CipherCellName"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:singleLine="true"
-                    android:ellipsize="end"
-                    android:text="Name" />
-                <TextView
-                    android:id="@+id/CipherCellSharedIcon"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="center_vertical"
-                    android:paddingLeft="5dp"
-                    android:singleLine="true"
-                    android:text="&#xf1e0;" />
-                <TextView
-                    android:id="@+id/CipherCellAttachmentsIcon"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="center_vertical"
-                    android:paddingLeft="5dp"
-                    android:singleLine="true"
-                    android:text="&#xf0c6;" />
-            </LinearLayout>
-            <TextView
-                android:id="@+id/CipherCellSubTitle"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:singleLine="true"
-                android:ellipsize="end"
-                android:text="SubTitle" />
-        </LinearLayout>
-        <Button
-            android:id="@+id/CipherCellButton"
-            android:layout_width="37dp"
-            android:layout_height="match_parent"
-            android:text="&#xe5d4;"
-            android:gravity="center"
-            android:padding="0dp"
-            android:background="@android:color/transparent" />
-    </LinearLayout>
-</RelativeLayout>
