+++ src/App/App.csproj
