+++ src/App/Controls/CipherViewCell/CipherViewCell.xaml
@@ -1,114 +1,102 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
-<ContentView xmlns="http://xamarin.com/schemas/2014/forms"
-          xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-          x:Class="Bit.App.Controls.CipherViewCell"
-          xmlns:controls="clr-namespace:Bit.App.Controls"
-          xmlns:u="clr-namespace:Bit.App.Utilities"
-          xmlns:ff="clr-namespace:FFImageLoading.Forms;assembly=FFImageLoading.Forms">
+<Grid xmlns="http://xamarin.com/schemas/2014/forms"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    x:Class="Bit.App.Controls.CipherViewCell"
+    xmlns:controls="clr-namespace:Bit.App.Controls"
+    xmlns:u="clr-namespace:Bit.App.Utilities"
+    xmlns:ff="clr-namespace:FFImageLoading.Forms;assembly=FFImageLoading.Forms"
+    x:Name="_grid"
+    StyleClass="list-row, list-row-platform"
+    RowSpacing="0"
+    ColumnSpacing="0"
+    x:DataType="controls:CipherViewCellViewModel">
 
-    <Grid 
-        x:Name="_grid"
-        StyleClass="list-row, list-row-platform"
-        RowSpacing="0"
-        ColumnSpacing="0"
-        x:DataType="controls:CipherViewCellViewModel">
+    <Grid.Resources>
+          <u:IconConverter x:Key="IconConverter"/>
+    </Grid.Resources>
 
-        <Grid.BindingContext>
-            <controls:CipherViewCellViewModel />
-        </Grid.BindingContext>
+    <Grid.BindingContext>
+        <controls:CipherViewCellViewModel />
+    </Grid.BindingContext>
 
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-        </Grid.RowDefinitions>
+    <Grid.RowDefinitions>
+        <RowDefinition Height="Auto" />
+    </Grid.RowDefinitions>
 
+    <Grid.ColumnDefinitions>
+        <ColumnDefinition Width="0.1*" />
+        <ColumnDefinition Width="0.75*" />
+        <ColumnDefinition Width="0.15*" />
+    </Grid.ColumnDefinitions>
+
+    <ff:CachedImage
+        x:Name="_image"
+        Grid.Column="0"
+        BitmapOptimizations="True"
+        ErrorPlaceholder="login.png"
+        LoadingPlaceholder="login.png"
+        HorizontalOptions="Center"
+        VerticalOptions="Center"
+        WidthRequest="22"
+        HeightRequest="22"
+        Source="{Binding Cipher, Converter={StaticResource IconConverter}}"
+        AutomationProperties.IsInAccessibleTree="False" />
+
+    <Grid RowSpacing="0" ColumnSpacing="0" Grid.Row="0" Grid.Column="1" VerticalOptions="Center">
         <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="40" />
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="Auto" />
             <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="60" />
         </Grid.ColumnDefinitions>
 
+        <Label 
+            LineBreakMode="TailTruncation"
+            Grid.Column="0"
+            Grid.Row="0"
+            Text="{Binding Cipher.Name}" />
+        <Label
+            LineBreakMode="TailTruncation"
+            Grid.Column="0"
+            Grid.Row="1"
+            Grid.ColumnSpan="3"
+            Text="{Binding Cipher.SubTitle}" />
         <controls:FaLabel
-            x:Name="_icon"
+            Grid.Column="1"
             Grid.Row="0"
-            Grid.Column="0"
-            HorizontalOptions="Center"
+            HorizontalOptions="Start"
             VerticalOptions="Center"
-            StyleClass="list-icon, list-icon-platform"
-            AutomationProperties.IsInAccessibleTree="False" />
-
-        <ff:CachedImage
-            x:Name="_image"
-            Grid.Row="0"
-            Grid.Column="0"
-            BitmapOptimizations="True"
-            ErrorPlaceholder="login.png"
-            HorizontalOptions="Center"
-            VerticalOptions="Center"
-            WidthRequest="22"
-            HeightRequest="22"
-            IsVisible="False"
-            AutomationProperties.IsInAccessibleTree="False" />
-
-        <Grid RowSpacing="0" ColumnSpacing="0" Grid.Row="0" Grid.Column="1" VerticalOptions="Center">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-            </Grid.RowDefinitions>
-
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
-
-            <Label 
-                LineBreakMode="TailTruncation"
-                Grid.Column="0"
-                Grid.Row="0"
-                StyleClass="list-title, list-title-platform"
-                Text="{Binding Cipher.Name, Mode=OneWay}" />
-            <Label
-                LineBreakMode="TailTruncation"
-                Grid.Column="0"
-                Grid.Row="1"
-                Grid.ColumnSpan="3"
-                StyleClass="list-subtitle, list-subtitle-platform"
-                Text="{Binding Cipher.SubTitle, Mode=OneWay}" />
-            <controls:FaLabel
-                Grid.Column="1"
-                Grid.Row="0"
-                HorizontalOptions="Start"
-                VerticalOptions="Center"
-                StyleClass="list-title-icon"
-                Margin="5, 0, 0, 0"
-                Text="&#xf1e0;"
-                IsVisible="{Binding Cipher.Shared, Mode=OneWay}"
-                AutomationProperties.IsInAccessibleTree="True"
-                AutomationProperties.Name="{u:I18n Shared}" />
-            <controls:FaLabel
-                Grid.Column="2"
-                Grid.Row="0"
-                HorizontalOptions="Start"
-                VerticalOptions="Center"
-                StyleClass="list-title-icon"
-                Margin="5, 0, 0, 0"
-                Text="&#xf0c6;"
-                IsVisible="{Binding Cipher.HasAttachments, Mode=OneWay}"
-                AutomationProperties.IsInAccessibleTree="True"
-                AutomationProperties.Name="{u:I18n Attachments}" />
-        </Grid>
-
-        <controls:MiButton
-            Grid.Row="0"
-            Grid.Column="2"
-            Text="&#xe5d3;"
-            StyleClass="list-row-button, list-row-button-platform, btn-disabled"
-            Clicked="MoreButton_Clicked"
-            VerticalOptions="CenterAndExpand"
-            HorizontalOptions="EndAndExpand"
+            StyleClass="list-title-icon"
+            Margin="5, 0, 0, 0"
+            Text="&#xf1e0;"
+            IsVisible="{Binding Cipher.Shared, Mode=OneTime}"
             AutomationProperties.IsInAccessibleTree="True"
-            AutomationProperties.Name="{u:I18n Options}" />
-
+            AutomationProperties.Name="{u:I18n Shared}" />
+        <controls:FaLabel
+            Grid.Column="2"
+            Grid.Row="0"
+            HorizontalOptions="Start"
+            VerticalOptions="Center"
+            StyleClass="list-title-icon"
+            Margin="5, 0, 0, 0"
+            Text="&#xf0c6;"
+            IsVisible="{Binding Cipher.HasAttachments, Mode=OneTime}"
+            AutomationProperties.IsInAccessibleTree="True"
+            AutomationProperties.Name="{u:I18n Attachments}" />
     </Grid>
 
-</ContentView>
+    <controls:MiButton
+        Grid.Row="0"
+        Grid.Column="2"
+        Text="&#xe5d3;"
+        StyleClass="list-row-button, list-row-button-platform, btn-disabled"
+        Clicked="MoreButton_Clicked"
+        VerticalOptions="CenterAndExpand"
+        HorizontalOptions="EndAndExpand"
+        AutomationProperties.IsInAccessibleTree="True"
+        AutomationProperties.Name="{u:I18n Options}" />
+
+    <BoxView StyleClass="list-item-separator-bottom, list-item-separator-bottom-platform"
+             Grid.ColumnSpan="3"
+             Grid.Row="1"
+             VerticalOptions="End" />
+</Grid>
