--- src/App/Controls/ExtendedCollectionView.cs
@@ -1,8 +1,0 @@
-﻿using Xamarin.Forms;
-
-namespace Bit.App.Controls
-{
-    public class ExtendedCollectionView : CollectionView
-    {
-    }
-}
