--- src/App/Controls/ExtendedListView.cs
@@ -1,12 +1,0 @@
-﻿using Xamarin.Forms;
-
-namespace Bit.App.Controls
-{
-    public class ExtendedListView : ListView
-    {
-        public ExtendedListView() { }
-
-        public ExtendedListView(ListViewCachingStrategy cachingStrategy)
-            : base(cachingStrategy) { }
-    }
-}
