+++ src/App/Pages/Settings/FoldersPage.xaml
@@ -32,27 +32,27 @@
                        VerticalOptions="CenterAndExpand"
                        HorizontalOptions="CenterAndExpand"
                        HorizontalTextAlignment="Center"></Label>
-                <ListView x:Name="_listView"
+                <CollectionView x:Name="_collectionView"
                           IsVisible="{Binding ShowNoData, Converter={StaticResource inverseBool}}"
                           ItemsSource="{Binding Folders}"
                           VerticalOptions="FillAndExpand"
-                          CachingStrategy="RecycleElement"
-                          ItemSelected="RowSelected"
+                          SelectionChanged="RowSelected"
                           StyleClass="list, list-platform">
-                    <ListView.ItemTemplate>
+                    <CollectionView.ItemTemplate>
                         <DataTemplate x:DataType="views:FolderView">
-                            <ViewCell>
-                                <StackLayout
-                                    StyleClass="list-row, list-row-platform"
-                                    Padding="10">
-                                    <Label LineBreakMode="TailTruncation"
-                                           StyleClass="list-title, list-title-platform"
-                                           Text="{Binding Name, Mode=OneWay}" />
-                                </StackLayout>
-                            </ViewCell>
+                            <StackLayout
+                                StyleClass="list-row, list-row-platform"
+                                Padding="10">
+                                <Label LineBreakMode="TailTruncation"
+                                        StyleClass="list-title, list-title-platform"
+                                        Text="{Binding Name, Mode=OneWay}" />
+                                <BoxView StyleClass="list-item-separator-bottom, list-item-separator-bottom-platform"
+                                         BackgroundColor="Black"
+                                         VerticalOptions="End" />
+                            </StackLayout>
                         </DataTemplate>
-                    </ListView.ItemTemplate>
-                </ListView>
+                    </CollectionView.ItemTemplate>
+                </CollectionView>
             </StackLayout>
         </ResourceDictionary>
     </ContentPage.Resources>
