+++ src/App/Pages/Settings/SettingsPage/SettingsPage.xaml
@@ -44,35 +44,31 @@
         </ResourceDictionary>
     </ContentPage.Resources>
 
-    <ListView
+    <CollectionView
         ItemsSource="{Binding GroupedItems}"
         VerticalOptions="FillAndExpand"
-        HasUnevenRows="True"
-        RowHeight="-1"
         ItemTemplate="{StaticResource listItemDataTemplateSelector}"
-        IsGroupingEnabled="True"
-        ItemSelected="RowSelected"
+        IsGrouped="True"
+        SelectionChanged="RowSelected"
         StyleClass="list, list-platform">
 
-        <ListView.GroupHeaderTemplate>
+        <CollectionView.GroupHeaderTemplate>
             <DataTemplate x:DataType="pages:SettingsPageListGroup">
-                <ViewCell>
-                    <StackLayout
-                        Padding="0" Spacing="0" VerticalOptions="FillAndExpand"
-                        StyleClass="list-row-header-container, list-row-header-container-platform">
-                        <BoxView
-                          StyleClass="list-section-separator-top, list-section-separator-top-platform"
-                          IsVisible="{Binding First, Converter={StaticResource inverseBool}}" />
-                        <StackLayout StyleClass="list-row-header, list-row-header-platform">
-                            <Label
-                                Text="{Binding Name}"
-                                StyleClass="list-header, list-header-platform" />
-                        </StackLayout>
-                        <BoxView StyleClass="list-section-separator-bottom, list-section-separator-bottom-platform" />
+                <StackLayout
+                    Padding="0" Spacing="0" VerticalOptions="FillAndExpand"
+                    StyleClass="list-row-header-container, list-row-header-container-platform">
+                    <BoxView
+                        StyleClass="list-section-separator-top, list-section-separator-top-platform"
+                        IsVisible="{Binding First, Converter={StaticResource inverseBool}}" />
+                    <StackLayout StyleClass="list-row-header, list-row-header-platform">
+                        <Label
+                            Text="{Binding Name}"
+                            StyleClass="list-header, list-header-platform" />
                     </StackLayout>
-                </ViewCell>
+                    <BoxView StyleClass="list-section-separator-bottom, list-section-separator-bottom-platform" />
+                </StackLayout>
             </DataTemplate>
-        </ListView.GroupHeaderTemplate>
-    </ListView>
+        </CollectionView.GroupHeaderTemplate>
+    </CollectionView>
 
 </pages:BaseContentPage>
