+++ src/App/Pages/Settings/SettingsPage/SettingsPage.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using Bit.App.Abstractions;
 using Bit.App.Resources;
 using Bit.Core.Utilities;
+using System.Linq;
 using System.Threading.Tasks;
 using Xamarin.Forms;
 
@@ -41,14 +42,14 @@ namespace Bit.App.Pages
             return base.OnBackButtonPressed();
         }
 
-        private async void RowSelected(object sender, SelectedItemChangedEventArgs e)
+        private async void RowSelected(object sender, SelectionChangedEventArgs e)
         {
-            ((ListView)sender).SelectedItem = null;
+            ((CollectionView)sender).SelectedItem = null;
             if (!DoOnce())
             {
                 return;
             }
-            if (!(e.SelectedItem is SettingsPageListItem item))
+            if (!(e.CurrentSelection?.FirstOrDefault() is SettingsPageListItem item))
             {
                 return;
             }
