+++ src/App/Styles/iOS.xaml
@@ -51,11 +51,14 @@
         <Setter Property="BackgroundColor"
                 Value="{StaticResource BackgroundColor}" />
     </Style>
-    <Style TargetType="ListView"
+    <Style TargetType="CollectionView"
            ApplyToDerivedTypes="True">
         <Setter Property="BackgroundColor"
                 Value="{StaticResource BackgroundColor}" />
-        <Setter Property="RefreshControlColor"
+    </Style>
+    <Style TargetType="RefreshView"
+           ApplyToDerivedTypes="True">
+        <Setter Property="RefreshColor"
                 Value="{StaticResource PrimaryColor}" />
     </Style>
     <Style TargetType="ActivityIndicator"
@@ -93,10 +96,10 @@
 
     <!-- List -->
 
-    <Style TargetType="ListView"
-           Class="list-platform"
+    <Style TargetType="BoxView"
+           Class="list-item-separator-bottom-platform"
            ApplyToDerivedTypes="True">
-        <Setter Property="SeparatorColor"
+        <Setter Property="BackgroundColor"
                 Value="{StaticResource ListItemBorderColor}" />
     </Style>
     <Style TargetType="StackLayout"
